feat: add CKEditorWrapper and implement it
