+++ src/app/blog/components/CKEditorWrapper.scss
@@ -1,0 +1,5 @@
+.ck.ck-editor {
+  .ck-editor__editable {
+    min-height: 200px;
+  }
+}
